Remove reveald.Request modifications
When generating continuation URLs, the original reveald.Request
should not be modified. If a copy is created which is operated on
then correct URLs for each continuation, such as aggregation, pagination,
and sorting, will be correct - and only include data from the current
search request (along with the continuation values).